# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Uptime Cairn API
  version: 1.0.0-draft
  summary: All-in-one open source uptime monitoring and reporting platform.
  description: |
    The complete public surface of Uptime Cairn. The dashboard is the first
    client of this API, never a privileged one — there are no internal
    endpoints the UI can reach that an API key cannot.

    ## Authentication

    Two schemes, either of which authenticates any non-public operation:

    - `ApiKeyAuth` — `Authorization: Bearer cairn_<key>`. Scoped, expiring,
      revocable. The scheme for automation, CI/CD, and Terraform.
    - `SessionCookie` — the browser session the dashboard uses after
      `POST /api/v1/auth/login`. Cookie-authenticated state-changing requests
      additionally require a CSRF token in `X-Cairn-CSRF-Token`.

    Operations tagged `Public` require no authentication: the status page read
    path, subscriber confirmation, and push-monitor ingest.

    ## Pagination

    Every list endpoint paginates by opaque cursor, keyed on `(updated_at, id)`,
    per ADR-004 — uniformly, with no exception for installs small enough that
    sending everything would "work". Offsets are unstable against rows that
    change while a user is paginating, and status changes constantly.

    Cursor responses carry no total count. A count for the current filter comes
    from `GET /api/v1/monitors/membership`, which a client polling for live-view
    freshness is already calling.

...

    ## Live updates
...
servers:
...
security:
  - ApiKeyAuth: []
  - SessionCookie: []
...
paths:
...
  /api/v1/groups:
    get:
      tags: [Groups]
      operationId: listGroups
      summary: List monitor groups
      x-cairn-phase: 1
      x-cairn-scopes: [groups:read]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/CollectionLimit"
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: A page of groups.
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Group" }
                  pagination: { $ref: "#/components/schemas/CursorPagination" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Groups]
      operationId: createGroup
      summary: Create a group
      x-cairn-phase: 1
      x-cairn-scopes: [groups:write]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GroupWrite" }
      responses:
        "201":
          description: Group created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Group" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
  /api/v1/tags:
    get:
      tags: [Tags]
      operationId: listTags
      summary: List tags
      x-cairn-phase: 1
      x-cairn-scopes: [tags:read]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/CollectionLimit"
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: A page of tags.
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Tag" }
                  pagination: { $ref: "#/components/schemas/CursorPagination" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Tags]
      operationId: createTag
      summary: Create a tag
      x-cairn-phase: 1
      x-cairn-scopes: [tags:write]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TagWrite" }
      responses:
        "201":
          description: Tag created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tag" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: cairn_<key>
      description: |
        A scoped API key: `Authorization: Bearer cairn_<key>`. Keys carry an explicit
        scope list, an optional expiry, and a last-used timestamp, and can be revoked
        at any time. Every operation names the scopes it requires in `x-cairn-scopes`.
    SessionCookie:
      type: apiKey
      in: cookie
      name: cairn_session
      description: |
        The browser session established by `POST /api/v1/auth/login`. The cookie is
        `HttpOnly`, `SameSite=Lax`, and `Secure` when served over HTTPS.
        State-changing requests authenticated this way must also carry a matching
        `X-Cairn-CSRF-Token` header.
  parameters:
    CollectionLimit:
      name: limit
      in: query
      description: |
        Maximum items to return. The server caps this; requesting more is clamped,
        not rejected.
      schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
    Search:
      name: q
      in: query
      description: Free-text search across the resource's human-readable fields.
      schema: { type: string, maxLength: 200 }
    Cursor:
      name: cursor
      in: query
      description: |
        Opaque cursor from a previous response's `pagination.next_cursor`. Cursors are
        not decodable, not sortable, and not guaranteed stable across releases.
      schema: { type: string }
...
  responses:
...
    Unauthorized:
      description: No credentials were supplied, or they were not valid.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    Conflict:
      description: The request conflicts with the current state — a duplicate slug, a referenced resource still in use.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ValidationFailed:
      description: |
        The request was well-formed but semantically invalid. `errors` names each
        offending field by JSON pointer.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ValidationProblem" }
...
  schemas:
...
    Problem:
      type: object
      description: An error, per RFC 9457.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: |
            A stable URI identifying the error kind. Clients branch on this, never on
            `title` or `detail`, both of which are prose and may be translated.
          examples: ["https://uptimecairn.dev/errors/monitor-not-found"]
        title:
          type: string
          description: A short, human-readable summary of the error kind.
        status:
          type: integer
          description: The HTTP status code, repeated for clients that only read the body.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: The request path that produced the error.
        request_id:
          type: string
          description: Correlates this response with the server log entry.
    ValidationProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            errors:
              type: array
              description: One entry per invalid field.
              items:
                type: object
                required: [pointer, code, message]
                properties:
                  pointer:
                    type: string
                    description: RFC 6901 JSON pointer to the offending field.
                    examples: ["/config/interval_seconds"]
                  code:
                    type: string
                    description: Stable machine-readable reason.
                    examples: ["below_minimum"]
                  message:
                    type: string
                    description: Human-readable explanation.
    CursorPagination:
      type: object
      description: |
        The only pagination model in this API, per ADR-004: every list view is a
        cursor-paginated query keyed on `(updated_at, id)`, applied uniformly
        regardless of how few rows an install has. Iterate until `next_cursor` is null.

        There is deliberately **no total count**. Computing one costs a full scan of
        the filtered set on every page fetch, which is the cost cursor pagination
        exists to avoid. A count for the current filter comes from
        `GET /api/v1/monitors/membership` instead, which returns it alongside the
        change signal the client is already polling.
      required: [next_cursor, has_more]
      properties:
        next_cursor:
          type: [string, "null"]
          description: Pass as `cursor` to fetch the next page. Null when the collection is exhausted.
        has_more: { type: boolean }
...
    MonitorStatus:
      type: string
      description: |
        `pending` covers both a monitor that has never reported and one that is retrying
        below its failure threshold — it has not yet earned a verdict either way.
      enum: [up, down, pending, paused, maintenance]
...
    Group:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, "null"], maxLength: 2000 }
        parent_group_id:
          type: [string, "null"]
          format: uuid
          description: Groups may nest one level deep in Phase 1. Cycles are rejected.
        monitor_count: { type: integer, minimum: 0, readOnly: true }
        status:
          allOf:
            - $ref: "#/components/schemas/MonitorStatus"
          readOnly: true
          description: Worst status among the group's monitors — down beats pending beats up.
        created_at: { type: string, format: date-time, readOnly: true }
        updated_at: { type: string, format: date-time, readOnly: true }
    GroupWrite:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, "null"], maxLength: 2000 }
        parent_group_id: { type: [string, "null"], format: uuid }
    Tag:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        name: { type: string, minLength: 1, maxLength: 100 }
        slug:
          type: string
          readOnly: true
          description: URL-safe form derived from the name, unique across the instance.
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
          default: "#6b7280"
          description: Hex colour used in the UI and on status pages.
        description: { type: [string, "null"], maxLength: 2000 }
        monitor_count: { type: integer, minimum: 0, readOnly: true }
        created_at: { type: string, format: date-time, readOnly: true }
        updated_at: { type: string, format: date-time, readOnly: true }
    TagWrite:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        color: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
        description: { type: [string, "null"], maxLength: 2000 }

    # ------------------------------------------------------------------------
    # Notification channels
    # ------------------------------------------------------------------------
...